Seperated modules to different files: - Sprites class and children (Player, Enenmy, Missle) to sprites.py - Game Class to game.y - Variable for game configuration to game_config.py
Additionally fixed a problem where some game config variables have not
been considred but static values have been used instead
(e.g. fieled_height and field_with)
