+++ game_config.py
@@ -1,0 +1,14 @@
+# Configuration variabales
+
+# Game Field
+field_width = 300
+field_height = 300
+
+# Spite Properties
+player_lives = 3
+player_speed_default = 4
+player_turn_rate = 24
+
+enemy_speed = 4
+
+bullet_speed = 20
